Added GSA analysis chart for sequential simulation.  The GSA analysis (see seq_gsa_anal.pdf) shows that the following variables have the most influence on performance difference between ladderQ and heap2tQ: 1. Events/agent, 2. Time step (delay), and 3. Fraction of self-events.
All other parameters do not have much influence on the difference.  So
the data narrows down the parameters from 9 to 3 for sequential
simulation comparisons.

+++ muse/Thesis/higirodj/parameter_analysis/plot_gsa_data.sh
@@ -8,11 +8,18 @@
 TMP_DIR="tmp/"
 mkdir -p "$TMP_DIR"
 
+# Ensure at least one input .csv file has been specified
+if [ $# -lt 1 ]; then
+    echo "Specify input CSV files to be aggregated and analyzed."
+    exit 2
+fi
+
 # File name for the full data file
 FULL_DATA="${TMP_DIR}/full_gsa.csv"
 rm -f "${FULL_DATA}"
 # Aggregate all the GSA data into a single file in TMP dir
-for dataFile in *.csv
+echo "Aggregating files: $* ..."
+for dataFile in $*
 do
     grep -v "\-\-" "$dataFile" >> "$FULL_DATA"
 done
